refactor: update import paths for ParticipantType to use new types directory
